TEAL-666: Implement DatabaseLoader class
The DatabaseLoader instance is initialised with a transactional SQLAlchemy connection to a MALTOPUFT DB instance. This allows database operations to be rolled back in the case where an error is encountered whi
le making many insertions into the target database.

The target MALTOPUFT DB connection details are currently inherited from ska-src-maltopuft-backend, but this should be made explicit to this module in the future.

The DatabaseLoader works by examining an input Pandas DataFrame and a 'target' dictionary (defined in targets.py). The target dictionary defines the SQLAlchemy model class of the insert, the prefix of relevant
 attribute columns in the DataFrame, the table primary key and the table foreign keys. With this information, the DatabaseLoader extracts the relevant DataFrame columns for the insert (including the primary and foreign keys). After the insertion is made, the database primary keys of each row are returned. These values are then used to update the values of the primary key column to the database primary keys. This means that, when inserting different tables *from the same DataFrame*, the primary and foreign keys are accurate.

Additionally, the DatabaseLoader keeps an internal map of local primary key and inserted primary key attributes.

Note that these values are not expected to be the same. This is because local primary keys are typically based on the DataFrame index (or integer row location) and therefore only identify unique rows in the pa
rticular ETL run. The database primary keys are always unique.

This internal map is kept to allow *local* foreign keys to be updated with database foreign keys across different local dataframe instances.