TEAL-666: Fix bug in beam_df merge
When the list of beams were exploded, the index was repeated. E.g. if the observation row at index 0 had a list of 12 beams, then the first 12 indexes in the exploded DataFrame were 0. This meant that a significant number of beam rows were dropped during the pd.merge operation on the table indexes. To resolve the issue, the index of the exploded DataFrame was reset to ensure that they were unique.