DAAC-215: Use tiling ra/dec as temporary observation ra/dec
The MeerTRAP run summary files don't have obvious ra/dec coordinates stored for an observation, so a temporary stand in is used to faciltate the known source search within a radius of an observation (i.e. to render known sources on the frontend).

The first observation tiling ra/dec coordinates are used as a stand-in/temporary ra/dec coordinate for the observation. Note that an observation can have many tilings, and *only the first is used here for the ra/dex coordinate*. A better strategy than using tiling coordinates as observation coordinates will be likeyly be required in the future. Even if the tiling coordinates are used in some capacity, then *all tilings* should be considered, and not just the first.