DAAC-215: Remove decimal places from ra/dec
A fairly 'dumb' strategy was being employed to validate ra and dec strings. The ska-src-maltopuft-backend repository defines the RaStr and DecStr types which each use a regex to parse the expected ra/dec coordinate pattern. The pattern assumes optional decimal places up to a fixed number, where the fixed number was chosen based on MeerTRAP precision. This (mostly) worked with MeerTRAP data without any issues, so no problems so far.

However, this strategy breaks down somewhat when messier data from other sources enters the picture. E.g. ra/dec coordinates in the ATNF pulsar catalogue sometimes have missing arc minutes, arc seconds, decimal places etc. and sometimes have longer than the fixed number of decimal places set on the backend validation side.

One solution to this is to set up log to trim decimal places and fill missing arc minutes/seconds and so on. However it's become clear that handling ra/dec coordinate strings in units of rms/dms respectively is troublesome, and a much easier way to do this is to use decimal degrees. If decimal degrees are used then the ra/dec values take continuous floating point values in the range [-90,90] and [0,180] respectively which makes them much easier to work with and validate.

So, despite attempting to use hms/dms strings so far because these are useful to pulsar/transient scientists, changes will be made to work with decimal degrees in the future. Therefore, this change just ensures that known source catalogue data matches the format expected by the backend in a simple way.