DAAC-216: Fix beam/candidate merge
The beam dataframe was created by finding the unique beams for the unique observations. This meant that there was no way to link another candidate from the same observation to a beam.

To resolve the issue, the beam DataFrame is merged with the DataFrame containing the candidates *by an outer join only on the observation_id*, rather than an inner join on observation_id and candidate which was being used previously.

Merging in this way means that candidate_x and candidate_y columns are created after the merge where candidate_x has n_cand unique candidates and candidate_y has n_obs unique candidates. To merge correctly, candidates_y is dropped and candidates_x is renamed to 'candidates'. Then, filtering on "beam == beam.number" (as before) gives n_cand rows along with the beam in which they were detected, which is the desired behaviour.