DAAC-251: Refactor handling of insert error
The DatabaseLoader.foreign_keys_map attribute has been updated to type list[tuple[int, int]]. This was required to support falling back to inserting observation rows because it allwos dataframe keys to be updated correctly if it first needs updating to a temporary key. This happens if the fetched database key already exists in the dataframe.

Previously, nested transactions were spawned from the sqlalchemy connection which was assigned to a variable when creating the engine. This meant that every DatabaseLoader method that communicated with the database required the `conn` parameter. The DatabaseLoader methods have been refactored to spawn the nested transactions from the self.conn attribute instead, which cleans up these function signatures.

The *_targets (formerly of type list[dict[str, Any]]) have been refactored to be instances of a TargetInformation dataclass instead. This allowed removal or a lot of type checks (e.g. target.get("...") is not None), and allows creating convenience functions to derive table names etc in a cleaner way. This is a much cleaner implementation and stops mypy complaining. Several other type checks/hints have been updated.

Any TargetDataFrameHandler methods which were effectively just thin wrappers around the pandas API have been removed, in favour of making the TargetDataFrameHandler.df attribute public and just using the pandas API directly.