DAAC-322: Only parse unique MeerTRAP run summary files
Identifies unique run summary files by computing the md5 hash of the file and storing the hashes in a hashmap. If the hash has been seen previously, the file is skipped. If the hash is not present in the hash map, then the file is parsed as before. As many run summary files are duplicates, this significantly reduces the number of files that need to be parsed. More importantly, this reduces the size of the DataFrame in which the parsed data is loaded making the transformation stage much more efficient wrt runtime and resource utilisation.

Because of these changes, functionality is added to relate candidates to the correct observation (and therefore beam_id). To achieve this, the get_candidate_beam_ids() uses the Polars DataFrame.join_asof method to search backwards in time to find the closest observation start time to the candidate detection time.