DAAC-464: Dyamically mount host volumes to launched containers
When dagster launches containers, the job launcher takes an image from the corresponding location. This means that volume mounts configured in the docker compose aren't carried across. In order to mount volumes, it would therefore be required to hardcode this into the Dockerfile and copy directories over on build. This is problematic because it would lead to large image sizes, increased build time, a lack of flexibility in environments and bake in lots of raw data that should probably not be shared into the image.

Unfortunately, the run_launcher in dagster.yaml doesn't accept environment variables to configure volume mounts from the host dynamically. To get around this, placeholder env vars have been added to dagster.yaml which correspond to the host path of volumes to be mounted. The entrypoints for services which are derived from this image have been updated to run envsubst to substitute the environment variables in the copy of dagster.yaml in the container.

Additionally, this commit removes COPY cfg/ from Dockerfile.atnf as this is only required by the Meertrap pipeline. The .env used in the docker-compose services has been replaced with a path to ./docker/.env, such that only one .env file is needed to bring the deployment up. Finally, /var/run/docker.sock is mounted to services by default, but this can be overriden by setting $DOCKER_SOCKET_PATH as a local environment variable in .env.